refactor: keep only one GatewayRuntime
