+++ jina/serve/runtimes/gateway/__init__.py
@@ -1,8 +1,16 @@
 import argparse
+import os
 from abc import ABC
 from typing import TYPE_CHECKING, Optional, Union
 
+from jina import __default_host__
+from jina.excepts import PortAlreadyUsed
+from jina.helper import is_port_free
+from jina.serve.gateway import BaseGateway
 from jina.serve.runtimes.asyncio import AsyncNewLoopRuntime
+from jina.serve.runtimes.gateway.grpc import GRPCGateway
+from jina.serve.runtimes.gateway.http import HTTPGateway
+from jina.serve.runtimes.gateway.websocket import WebSocketGateway
 
 if TYPE_CHECKING:
     import asyncio
@@ -28,3 +36,70 @@ class GatewayRuntime(AsyncNewLoopRuntime, ABC):
         if self.timeout_send:
             self.timeout_send /= 1e3  # convert ms to seconds
         super().__init__(args, cancel_event, **kwargs)
+
+    async def async_setup(self):
+        """
+        The async method setup the runtime.
+
+        Setup the uvicorn server.
+        """
+        if not self.args.proxy and os.name != 'nt':
+            os.unsetenv('http_proxy')
+            os.unsetenv('https_proxy')
+
+        if not (is_port_free(__default_host__, self.args.port)):
+            raise PortAlreadyUsed(f'port:{self.args.port}')
+
+        self.gateway = BaseGateway.load_config(
+            self.args.uses,
+            uses_with=dict(
+                name=self.name,
+                grpc_server_options=self.args.grpc_server_options,
+                port=self.args.port,
+                title=self.args.title,
+                description=self.args.description,
+                no_debug_endpoints=self.args.no_debug_endpoints,
+                no_crud_endpoints=self.args.no_crud_endpoints,
+                expose_endpoints=self.args.expose_endpoints,
+                expose_graphql_endpoint=self.args.expose_graphql_endpoint,
+                cors=self.args.cors,
+                ssl_keyfile=self.args.ssl_keyfile,
+                ssl_certfile=self.args.ssl_certfile,
+                uvicorn_kwargs=self.args.uvicorn_kwargs,
+            ),
+            uses_metas={},
+            runtime_args={  # these are not parsed to the yaml config file but are pass directly during init
+                'name': self.args.name,
+            },
+            py_modules=self.args.py_modules,
+            extra_search_paths=self.args.extra_search_paths,
+        )
+
+        self.gateway.set_streamer(
+            args=self.args,
+            timeout_send=self.timeout_send,
+            metrics_registry=self.metrics_registry,
+            runtime_name=self.args.name,
+        )
+        await self.gateway.setup_server()
+
+    async def _wait_for_cancel(self):
+        """Do NOT override this method when inheriting from :class:`GatewayPod`"""
+        # handle terminate signals
+        while not self.is_cancel.is_set() and not self.gateway.should_exit:
+            await asyncio.sleep(0.1)
+
+        await self.async_cancel()
+
+    async def async_teardown(self):
+        """Shutdown the server."""
+        await self.gateway.teardown()
+        await self.async_cancel()
+
+    async def async_cancel(self):
+        """Stop the server."""
+        await self.gateway.stop_server()
+
+    async def async_run_forever(self):
+        """Running method of the server."""
+        await self.gateway.run_server()
